Add support for the highlight plugin from pages Refactor cfmedata to be in the "fixtures" package. Automatically search for cfme_pages directory. Punt if it's not found. Add a requirements.txt to the tests repo
