+++ requirements.txt
@@ -1,0 +1,12 @@
+# This pulls in all the libraries needed to run Selenium tests
+# on Mozilla WebQA projects
+
+PyYAML==3.10
+UnittestZero
+py==1.4.12
+pytest==2.3.4
+pytest-mozwebqa==1.0
+requests==0.13.3
+selenium
+paramiko
+suds
